fix FindLib makefile to make it buildable standalone
Copy the "features" CFLAGS portion and some missing defines from the main makefile to fix the
standalone build of FindLib.

Change the definition of OBJ_DIR to match the main makefile.

FindLib may be useful as a standalone library which could be packaged separately and TraX made
dependendent on it.